move getDataFromMovieDB to the feature
